Renaming resque-lock-timeout with active_job_lock.
