+++ lib/active_job_lock.rb
@@ -1,0 +1,5 @@
+require 'active_job_lock/core'
+
+module ActiveJobLock
+  VERSION = '0.1.0'.freeze
+end
