--- lib/resque-lock-timeout.rb
@@ -1,2 +1,0 @@
-require 'resque'
-require 'resque/plugins/lock_timeout'
